+++ app/hotels/views.py
@@ -2,8 +2,9 @@ from datetime import date, datetime, timedelta
 from fastapi import APIRouter, Depends, Query
 
 from app.auth.dependes import get_current_user
+from app.expection import CannotBookHotelForLongPeriod, DateFromCannotBeAfterDateTo, HotelAlreadyException, NotFoundHotelsException
 from app.hotels.repositories import HotelsRepository
-from app.hotels.shemas import SHotels, SHotelsInfo
+from app.hotels.shemas import SHotels, SHotelsInfo, SHotelsLocation
 from app.users.models import Users
 
 
@@ -13,7 +14,7 @@ router = APIRouter(
 )
 
 @router.get('/id/{hotel_id}')
-async def get_hotel_by_id(hotel_id: int):
+async def get_hotel_by_id(hotel_id: int) -> SHotels:
     return await HotelsRepository.find_by_id(hotel_id)
 
 @router.get('/')
@@ -25,25 +26,34 @@ async def get_hotels_by_location(
     location: str,
     date_from: date = Query(description=f"{datetime.now().date()}"),
     date_to: date = Query(description=f"{(datetime.now() + timedelta(days=30)).date()}"),
-    ):
+    ) -> list[SHotelsLocation]:
     if date_from > date_to:
-        raise ValueError("date_from must be less than or equal to date_to")
+        raise DateFromCannotBeAfterDateTo
     if (date_to - date_from).days > 31:
-        raise ValueError("Maximum allowed duration is 31 days")
+        raise CannotBookHotelForLongPeriod
     return await HotelsRepository.find_all(location, date_from, date_to)
     
 @router.post('/')
 async def add_hotel(SHotels: SHotels, users: Users = Depends(get_current_user)):
+    hotels = HotelsRepository.find_one_or_none(name=SHotels.name)
+    if not hotels:
+        raise HotelAlreadyException
     await HotelsRepository.add(**SHotels.model_dump())
-    return {"detail": "Hotel added successfully"}
+    return {"detail": "Отель успешно добавлен"}
 
-@router.patch('/{hotel_id}')
+@router.patch('/change/{hotel_id}')
 async def update_hotel(hotel_id: int, SHotels: SHotels, users: Users = Depends(get_current_user)):
+    hotels = HotelsRepository.find_one_or_none(id=hotel_id)
+    if not hotels:
+        raise NotFoundHotelsException
     await HotelsRepository.update(hotel_id, **SHotels.model_dump())
-    return {"detail": "Hotel updated successfully"}
+    return {"detail": "Отель успешно обновлён"}
     
 @router.delete('/{hotel_id}')
 async def delete_hotel(hotel_id: int, users: Users = Depends(get_current_user)):
+    hotels = HotelsRepository.find_one_or_none(id=hotel_id)
+    if not hotels:
+        raise NotFoundHotelsException
     await HotelsRepository.delete(hotel_id)
-    return {"detail": "Hotel deleted successfully"}
+    return {"detail": "Отель успешно удалён"}
